🏗️ `webpm`: improve backend blueprint
