+++ src/youwol/backends/webpm/app/__init__.py
@@ -1,0 +1,4 @@
+# relative
+from .dependencies import lifespan
+from .middlewares import COEPHeaderMiddleware, CORSMiddleware
+from .paths import router
