+++ src/youwol/backends/webpm/app/middlewares/__init__.py
@@ -1,0 +1,4 @@
+# third parties
+from starlette.middleware.cors import CORSMiddleware
+
+from .coep import COEPHeaderMiddleware
