+++ src/youwol/backends/webpm/app/middlewares/coep.py
@@ -1,0 +1,13 @@
+# third parties
+from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
+from starlette.requests import Request
+from starlette.responses import Response
+
+
+class COEPHeaderMiddleware(BaseHTTPMiddleware):
+    async def dispatch(
+        self, request: Request, call_next: RequestResponseEndpoint
+    ) -> Response:
+        response = await call_next(request)
+        response.headers["cross-origin-resource-policy"] = "cross-origin"
+        return response
