+++ src/youwol/backends/webpm/app/paths/__init__.py
@@ -1,0 +1,3 @@
+# relative
+from .cdn_client import *
+from .root import *
