+++ src/youwol/backends/webpm/deployment/__init__.py
@@ -1,0 +1,3 @@
+# relative
+from .configuration import Configuration, ConfigurationFactory
+from .deployment import WebpmDeployment
